:hammer:Refact: develop mobile version css #1 #2 #5 #15
