--- mobile-index.html
@@ -1,72 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <!-- CSS -->
-    <link rel="stylesheet" href="static/css/mobile-index.css" />
-    <title>ChocoTheCoo</title>
-</head>
-
-<!-- header -->
-<header></header>
-
-<!-- body -->
-
-<body>
-    <div class="section">
-        <div class="row">
-            <div class="col-md-4">
-                <div class="collapse" id="navbarToggleExternalContent">
-                    <div class="p-4">
-                        <div class="menu" id="categorymenu">
-
-                        </div>
-                    </div>
-                </div>
-                <nav class="navbar navbar-light">
-                    <div class="container-fluid">
-                        <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                            data-bs-target="#navbarToggleExternalContent" aria-controls="navbarToggleExternalContent"
-                            aria-expanded="false" aria-label="Toggle navigation">
-                            <span class="navbar-toggler-icon"></span>
-                        </button>
-                    </div>
-                </nav>
-                <div class="search">
-                    <input type="text" id="search-keyword" placeholder="검색어 입력">
-                    <button type="button" class="btn btn-light" id="search-btn">🔍</button>
-                </div>
-            </div>
-            <div class="col" style="text-align:center; align-items: center;">
-                <img src="static/images/초코구독배너.gif" id="chocobanner">
-                <div class="row row-cols-1 row-cols-md-3 g-4" id="product-content">
-                    <!--상품정보-->
-                </div>
-                <div id="product-buttons">
-                </div>
-            </div>
-        </div>
-    </div>
-</body>
-<!-- footer -->
-<footer></footer>
-<!-- script -->
-<script src="/static/js/api.js" type="module"></script>
-<script src="/static/js/loader.js" type="module"></script>
-<script src="/static/js/mobile-index.js" type="module"></script>
-<script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.2/dist/umd/popper.min.js"
-    integrity="sha384-IQsoLXl5PILFhosVNubq5LC7Qb9DXgDA9i+tQ8Zj3iwWAwPtgFTxbJ8NT4GN1R8p"
-    crossorigin="anonymous"></script>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.min.js"
-    integrity="sha384-cVKIPhGWiC2Al4u+LWgxfKTRIcfu0JTxR+EQDz/bgldoEyl4H0zUF0QKbrJ0EcQF"
-    crossorigin="anonymous"></script>
-
-</html>
