--- mobile-mypage.html
@@ -1,153 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
-        crossorigin="anonymous"></script>
-    <!-- CSS -->
-    <link rel="stylesheet" href="static/css/mobile-mypage.css" />
-    <title>ChocoTheCoo</title>
-</head>
-
-<!-- header -->
-<header></header>
-
-<!-- body -->
-
-<body>
-    <div class="section">
-        <div class="row">
-            <div class="mypage-intro">
-                <p id="myreward">나의 리워드</p>
-                <div class="button-sets" style="display: inline; margin-left:50px;">
-                    <button class="Attendance" id="Attendance">
-                        출석인증
-                    </button>
-                    <button class="purchase-list" onclick="location.href = 'bill.html'">
-                        구매내역
-                    </button>
-                    <button class="review-see" onclick="location.href = 'myreview.html'">
-                        리뷰보기
-                    </button>
-                </div>
-                <img class="help-text3" src="/static/images/helpbtn.png" data-bs-toggle="tooltip"
-                    data-bs-placement="bottom" data-bs-html="true" data-bs-title="출석인증👉출석포인트지급<br/>(하루에 한 번)<br/>
-                        구매내역👉주문내역확인 및 구매확정, 리뷰작성!!<br/>리뷰내역👉리뷰확인 및 수정<br/>내 정보 수정👉프로필 수정">
-            </div>
-            <div class="point-all-view" style="display:flex;">
-                <p id="month-total"></p>
-                <p id="total-reward"></p>
-            </div>
-            <hr class="line">
-            <div class="point-calender" style="display: flex;">
-                <!--달력-->
-                <div class="calendar-section">
-                    <img class="help-text" src="/static/images/helpbtn.png" data-bs-toggle="tooltip"
-                        data-bs-placement="bottom" data-bs-title="날짜를 클릭하면 해당 포인트 정보를 확인할 수 있어요!">
-                    <table class="Calendar">
-                        <thead>
-                            <tr>
-                                <td id="prevCalendar" style="cursor:pointer;">&#60;</td>
-                                <td colspan="5">
-                                    <span id="calYear"></span>년
-                                    <span id="calMonth"></span>월
-                                </td>
-                                <td id="nextCalendar" style="cursor:pointer;">&#62;</td>
-                            </tr>
-                            <tr>
-                                <td>일</td>
-                                <td>월</td>
-                                <td>화</td>
-                                <td>수</td>
-                                <td>목</td>
-                                <td>금</td>
-                                <td>토</td>
-                            </tr>
-                        </thead>
-                        <tbody>
-                        </tbody>
-                    </table>
-                </div>
-                <div class="card-list">
-                    <div class="card-points" id="point-box">
-                        <p id="checkpoint"> 선택한 날짜의 포인트내역 확인하기</p>
-                        <div class="points-list" id="points-list">
-                            <!--포인트내역-->
-                        </div>
-                        <div id="point-buttons"></div>
-                    </div>
-                    <div class="statistic-list" id="statistic-list">
-                    </div>
-                </div>
-            </div>
-            <div class="top-category">
-                <nav class="product-nav">
-                    <a class="wishfont" id="wishfont" style='cursor:pointer'>위시리스트</a>
-                    <a class="bookmarkmarket" id="bookmarkmarket" style='cursor:pointer'>북마크한 마켓</a>
-                    <img class="help-text2" src="/static/images/helpbtn.png" data-bs-toggle="tooltip"
-                        data-bs-placement="right" data-bs-title="위시리스트 제품을 누르면 제품상세페이지로 이동합니다.">
-                </nav>
-                <div class="nav-underline"></div>
-            </div>
-            <hr class="line">
-            <div class="row">
-                <div class="col-md-12">
-                    <div id="my-wish-box" class="option-container">
-                        <div class="row row-cols-1 row-cols-md-3 g-4" id="my-wish-list">
-                            <!--위시리스트-->
-                        </div>
-                        <div class="my-wish-none"></div>
-                        <div id="wish-buttons"></div>
-                    </div>
-                    <div id="bookmarkList" class="option-container" style="display: none;">
-                        <!-- 팔로우 리스트 -->
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
-    </div>
-    <!--프로필부분-->
-    <div class="profile-section">
-        <div class="profile-card">
-            <div class="bigbox">
-                <div class="box">
-                    <img id="user-image" class="profile" src="static/images/default.png" />
-                </div>
-            </div>
-            <div class="profile-text">
-                <p>Name: <span id="user-name"></span></p>
-                <p>Email: <span id="user-email"></span></p>
-                <p>소개글: <span id="user-intro"></span></p>
-                <p>찜한 상품 개수: <span id="user-wish"></span></p>
-                <p>point: <span id="user-point"></span></p>
-                <p id="user-edit" onclick="location.href='/user_detail_page.html';">내 정보 수정</p>
-            </div>
-        </div>
-        <div class="subscription-section">
-            <button class="sub-button" id="sub-button"></button>
-            <div class="subscription-card" id="subscription-card">
-                <!--구독정보-->
-            </div>
-        </div>
-    </div>
-    </div>
-    </div>
-    </div>
-    </div>
-</body>
-<!-- footer -->
-<footer></footer>
-<!-- script -->
-<script src="/static/js/api.js" type="module"></script>
-<script src="/static/js/mypage.js" type="module"></script>
-<script src="/static/js/loader.js" type="module"></script>
-
-</html>
