+++ mypage.html
@@ -132,8 +132,8 @@
                         </div>
                     </div>
                 </div>
-                <button class="sub-button" id="sub-button"></button>
                 <div class="subscription-section">
+                    <button class="sub-button" id="sub-button"></button>
                     <div class="subscription-card" id="subscription-card">
                         <!--구독정보-->
                     </div>
