+++ myreview.html
@@ -8,6 +8,9 @@
     <!-- Bootstrap -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4"
+        crossorigin="anonymous"></script>
     <!-- CSS -->
     <link rel="stylesheet" href="static/css/myreview.css" />
     <title>ChocoTheCoo</title>
@@ -64,8 +67,8 @@
                         </div>
                     </div>
                 </div>
+                <button class="sub-button" id="sub-button"></button>
                 <div class="subscription-section">
-                    <button class="sub-button" id="sub-button"></button>
                     <div class="subscription-card" id="subscription-card">
                         <!--구독정보-->
                     </div>
