+++ static/css/bill.css
@@ -12,6 +12,10 @@
     font-style: normal;
 }
 
+body {
+    background-color: #C4A492;
+}
+
 button {
     height: 3vw;
     font-size: 1vw;
@@ -214,7 +218,31 @@ button {
 }
 
 
-@media screen and (max-width:1080px) {
+@media screen and (max-width: 800px) {
+
+    .pd-info-thumb {
+        object-fit: cover;
+        width: 20vw;
+        height: auto;
+        box-shadow: 5px 5px 5px rgba(0, 0, 0, 0.5);
+        border-radius: 4px;
+    }
+
+
+    .pd-info {
+        width: 100%;
+        display: flex;
+        font-family: 'S-CoreDream-3Light';
+        font-size: 2vw;
+        color: aliceblue;
+        margin-bottom: 1em;
+    }
+
+    #billLists {
+        display: flex;
+        flex-direction: column;
+        width: 80%;
+    }
 
     #sectionBox {
         display: flex;
@@ -237,4 +265,58 @@ button {
     #billLists {
         width: 100%;
     }
+
+    .profile-section {
+        width: 100%;
+        justify-content: center;
+        align-items: center;
+    }
+
+    .sub-button {
+        width: 80%;
+        margin-left: 25px;
+    }
+
+    .subscription-section {
+        width: 80%;
+        margin-right: 15px;
+    }
+
+    .profile {
+        width: 23vw;
+    }
+
+    .profile-card {
+        font-family: 'S-CoreDream-3Light';
+        font-size: 1.5vw;
+        width: 80%;
+        height: auto;
+        text-align: center;
+        margin-top: 10px;
+        margin-bottom: 10px;
+        margin-left: 20px;
+        border: 2px solid white;
+        border-radius: 20px;
+        color: white;
+    }
+
+    .subscription-card {
+        margin-left: 20px;
+        height: auto;
+        width: 100%;
+        padding: 20px;
+        text-align: center;
+        margin-bottom: 20px;
+        border: 2px solid white;
+        border-radius: 20px;
+        font-family: 'S-CoreDream-3Light';
+        font-size: 1.5vw;
+        color: white;
+        font-weight: bold;
+    }
+
+    .line {
+        width: 90%;
+    }
+
 }
