+++ static/css/index.css
@@ -256,3 +256,128 @@ button {
 .form-check {
   text-align: left;
 }
+
+
+.section-mobile {
+  display: none;
+}
+
+@media screen and (max-width: 768px) {
+  .category {
+    display: none;
+  }
+
+  .col-md-6 {
+    width: 100%;
+  }
+
+  .right-side {
+    display: none;
+  }
+
+  .section>#product-content {
+    display: block;
+  }
+
+  .section-mobile {
+    display: block;
+    background-color: #C4A492 !important;
+  }
+
+  body {
+    background-color: #C4A492;
+  }
+
+  .img {
+    border-radius: 20px;
+  }
+
+  .search-mobile {
+    display: flex;
+    position: relative;
+    padding-left: 20px;
+    align-items: center;
+    font-size: 20px;
+    margin-top: 10px;
+  }
+
+  #search-btn-mobile {
+    background-color: transparent;
+    border: none;
+    height: 40px;
+  }
+
+
+  input {
+    width: 100%;
+    border: 1px solid #bbb;
+    border-radius: 8px;
+    padding: 10px 12px;
+    font-size: 14px;
+    height: 40px;
+  }
+
+  .menu {
+    margin-left: 30px;
+    width: 30vw;
+    font-family: 'TheJamsil5Bold';
+    font-size: 4vw;
+    font-weight: bold;
+  }
+
+  .navbar {
+    width: 100%;
+  }
+
+  .section {
+    background-color: #C4A492;
+    height: auto;
+  }
+
+  #product-content {
+    margin-top: 10px;
+    padding: 20px;
+    text-align: center;
+  }
+
+  #product-buttons {
+    position: relative;
+    padding: 1rem 0;
+    display: flex;
+    justify-content: center;
+  }
+
+  #chocobanner {
+    position: relative;
+    height: auto;
+    width: 80%;
+    margin-top: 20px;
+    cursor: pointer;
+  }
+
+  .card {
+    font-family: 'S-CoreDream-3Light';
+    height: 40vw;
+    min-height: max-content;
+    width: auto;
+    font-size: 1vw;
+    overflow: auto;
+  }
+
+  button {
+    height: 4vh;
+    font-size: 1vw;
+    background-color: rgba(109, 86, 68, 0.822);
+    color: white;
+    border-radius: 10px;
+  }
+
+  #previous-buttons {
+    margin-right: 10px;
+  }
+
+  #next-buttons {
+    margin-left: 10px;
+  }
+
+}
