--- static/css/mobile-index.css
@@ -1,154 +1,0 @@
-@font-face {
-  font-family: '국립박물관문화재단클래식B';
-  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_twelve@1.0/국립박물관문화재단클래식B.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'S-CoreDream-3Light';
-  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_six@1.2/S-CoreDream-3Light.woff') format('woff');
-  font-weight: normal;
-  font-style: normal;
-}
-
-@font-face {
-  font-family: 'TheJamsil5Bold';
-  src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2302_01@1.0/TheJamsil5Bold.woff2') format('woff2');
-  font-weight: 700;
-  font-style: normal;
-}
-
-.section {
-  background-color: #C4A492!important;
-}
-
-body {
-  background-color: #C4A492;
-}
-
-.img {
-  border-radius: 20px;
-}
-
-.search {
-  display: flex;
-  position: relative;
-  padding-left: 20px;
-  align-items: center;
-  font-size: 20px;
-  margin-top: 10px;
-}
-
-#search-btn {
-  background-color: transparent;
-  border: none;
-  height: 40px;
-}
-
-
-input {
-  width: 100%;
-  border: 1px solid #bbb;
-  border-radius: 8px;
-  padding: 10px 12px;
-  font-size: 14px;
-  height: 40px;
-}
-
-.menu {
-  margin-left: 30px;
-  width: 30vw;
-  font-family: 'TheJamsil5Bold';
-  font-size: 4vw;
-  font-weight: bold;
-}
-
-.navbar {
-  width: 100%;
-}
-
-.section {
-  background-color: #C4A492;
-  height: auto;
-}
-
-#product-content {
-  margin-top: 10px;
-  padding: 20px;
-  text-align: center;
-}
-
-#product-buttons {
-  position: relative;
-  padding: 1rem 0;
-  display: flex;
-  justify-content: center;
-}
-
-#chocobanner {
-  position: relative;
-  height: auto;
-  width: 80%;
-  margin-top: 20px;
-  cursor: pointer;
-}
-
-.image-card {
-  height: 90px;
-  min-height: 90px;
-  max-height: 90px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.card-img-top {
-  height: fit-content;
-  object-fit: cover;
-  max-width: 100%;
-}
-
-.card {
-  font-family: 'S-CoreDream-3Light';
-  height: 200px;
-  min-height: max-content;
-  width: auto;
-  font-size: 1vw;
-  overflow: auto;
-}
-
-.card::-webkit-scrollbar {
-  display: none;
-}
-
-.carddesc {
-  height: fit-content;
-  object-fit: cover;
-  align-items: center;
-}
-
-.card-title {
-  font-weight: bold;
-  font-size: 1.5vw;
-}
-
-.card-body {
-  height: 50px;
-}
-
-button {
-  height: 4vh;
-  font-size: 1vw;
-  background-color: rgba(109, 86, 68, 0.822);
-  color: white;
-  border-radius: 10px;
-}
-
-#previous-buttons {
-  margin-right: 10px;
-}
-
-#next-buttons {
-  margin-left: 10px;
-}
