+++ static/css/mypage.css
@@ -412,3 +412,384 @@ td {
   background-color: black;
   color: white;
 }
+
+
+@media screen and (max-width: 768px) {
+  .col-md-4 {
+    justify-content: center;
+    align-items: center;
+  }
+
+  .section {
+    background-color: #C4A492;
+    height: auto;
+  }
+
+  .mypage-intro {
+    display: flex;
+    margin-top: 20px;
+  }
+
+  #myreward {
+    font-family: '국립박물관문화재단클래식B';
+    font-size: 4vw;
+    font-weight: bold;
+    padding-left: 40px;
+    color: white;
+  }
+
+  .card-list {
+    width: auto;
+  }
+
+  .statistic-list {
+    width: 40vw;
+    background-color: white;
+    height: fit-content;
+    margin-left: 20px;
+    margin-top: 20px;
+    padding: 10px;
+    text-align: center;
+    font-family: '국립박물관문화재단클래식B';
+    font-size: 2vw;
+    font-weight: bold;
+    padding-bottom: 10px;
+    color: rgba(100, 81, 66, 0.822);
+  }
+
+  #checkpoint {
+    font-family: '국립박물관문화재단클래식B';
+  }
+
+  .card-points {
+    width: 40vw;
+    background-color: white;
+    height: fit-content;
+    margin-left: 20px;
+    margin-top: 20px;
+    padding: 10px;
+    text-align: center;
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.5vw;
+    font-weight: bold;
+    padding-bottom: 10px;
+    color: rgba(100, 81, 66, 0.822);
+  }
+
+  .points-list {
+    font-family: 'S-CoreDream-3Light';
+    text-align: left;
+    font-size: 1.35vw;
+    font-weight: lighter;
+    padding-bottom: 3px;
+    margin-left: 20px;
+    color: rgba(100, 81, 66, 0.822);
+  }
+
+  .point-statistic {
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.2vw;
+    margin-bottom: 5px;
+    margin-top: 5px;
+  }
+
+  .help-text {
+    position: absolute;
+    object-fit: cover;
+    width: 20px;
+    height: 20px;
+    left: 30px;
+    z-index: 1;
+  }
+
+  .help-text2 {
+    position: absolute;
+    object-fit: cover;
+    width: 20px;
+    height: 20px;
+    right: -40px;
+    top: 20px;
+    z-index: 1;
+  }
+
+  .help-text3 {
+    position: absolute;
+    object-fit: cover;
+    width: 20px;
+    height: 20px;
+    left: 10px;
+    z-index: 1;
+  }
+
+  /* 크기변경 */
+  .tooltip>.tooltip-inner {
+    max-width: 400px;
+  }
+
+  /* 위치
+  .tooltip {
+    inset: auto auto 100px 100px;
+  } */
+
+  #month-total {
+    font-family: 'S-CoreDream-3Light';
+    font-size: 30px;
+    font-weight: bold;
+    margin-left: 40px;
+    color: white;
+  }
+
+  #total-reward {
+    font-family: 'S-CoreDream-3Light';
+    font-size: 30px;
+    font-weight: bold;
+    margin-left: 50px;
+    color: white;
+  }
+
+  #totalpoint {
+    margin-top: 10px;
+    font-weight: bold;
+  }
+
+  td {
+    width: 90px;
+    height: auto;
+  }
+
+  .Calendar {
+    text-align: center;
+    width: 33vw;
+    height: fit-content;
+    font-size: 1vw;
+  }
+
+
+  .line {
+    border: 2px solid rgba(255, 255, 255, 0.729);
+    width: 85%;
+    margin-left: 30px;
+  }
+
+  .calendar-section {
+    background-color: white;
+    margin-left: 30px;
+    margin-bottom: 20px;
+    margin-top: 20px;
+    width: 40vw;
+    height: fit-content;
+    border-radius: 5px;
+    box-shadow: 0px 35px 20px -30px rgba(0, 0, 0, 0.3);
+    padding: 20px;
+    display: flex;
+    justify-content: center;
+  }
+
+  .profile-card {
+    width: 100%;
+    margin: 0;
+  }
+
+  .profile-section {
+    width: 100%;
+    justify-content: center;
+    align-items: center;
+    margin-left: 20px;
+  }
+
+  .sub-button {
+    width: 100%;
+    margin-left: 30px;
+  }
+
+  .subscription-section {
+    width: 80%;
+    margin-left: 20px;
+  }
+
+  .profile {
+    width: 23vw;
+  }
+
+  .profile-card {
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.5vw;
+    width: 80%;
+    height: auto;
+    text-align: center;
+    margin-top: 30px;
+    margin-left: 30px;
+    border: 2px solid white;
+    border-radius: 20px;
+    color: white;
+  }
+
+  .subscription-card {
+    margin-left: 30px;
+    height: auto;
+    width: 100%;
+    padding: 20px;
+    text-align: center;
+    margin-bottom: 20px;
+    border: 2px solid white;
+    border-radius: 20px;
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.5vw;
+    color: white;
+    font-weight: bold;
+  }
+
+
+  /* section calendar 끝 */
+
+  /*프로필 기본 사진*/
+  .bigbox {
+    display: flex;
+    justify-content: center;
+    margin-bottom: 30px;
+  }
+
+  .box {
+    margin-top: 20px;
+  }
+
+  #user-edit {
+    cursor: pointer;
+  }
+
+  #wishitem {
+    padding: 20px;
+    text-align: center;
+    width: 60vw;
+    height: auto;
+    margin-left: 20vw;
+    border: 2px solid white;
+    border-radius: 20px;
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.5vw;
+    align-items: center;
+    color: black;
+  }
+
+  #my-wish-list {
+    font-family: 'S-CoreDream-3Light';
+    font-size: 1.5vw;
+    font-weight: bold;
+    color: white;
+
+  }
+
+  .my-wish-none {
+    margin-left: 30px;
+    color: white;
+    font-family: 'S-CoreDream-3Light';
+  }
+
+  .wishfont {
+    font-family: 'S-CoreDream-3Light';
+    font-weight: bold;
+    padding-left: 40px;
+    margin-top: 20px;
+    font-size: 2vw;
+    color: white;
+  }
+
+  .bookmarkmarket {
+    font-family: 'S-CoreDream-3Light';
+    font-weight: bold;
+    padding-left: 40px;
+    margin-top: 20px;
+    font-size: 2vw;
+    color: white;
+  }
+
+  .product-nav {
+    justify-content: center;
+    position: relative;
+    display: flex;
+    width: auto;
+    height: 50px;
+  }
+
+  .wishimage {
+    width: 17vw;
+    height: 10vw;
+  }
+
+  .wishname {
+    margin-top: 10px;
+    margin-bottom: 10px;
+  }
+
+  .subscrip-view {
+    margin-bottom: 1.5vw;
+  }
+
+  #bookmarkList {
+    height: auto;
+  }
+
+  .seller-card-list {
+    height: 15rem;
+    width: 100%;
+    padding: 2% 2% 2% 2%;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-left: 20px;
+  }
+
+  .seller-card-box {
+    width: 70%;
+    height: 100%;
+    margin-right: 5%;
+    border-radius: 2rem;
+    display: flex;
+    border: 2px solid rgba(255, 255, 255, 0.729);
+
+  }
+
+  .seller-img-box {
+    height: 100%;
+    width: 40%;
+    background-size: 100% 100%;
+    background-color: transparent;
+    background-position: center;
+    background-repeat: no-repeat;
+    border-radius: 50%;
+  }
+
+  .seller-information-box {
+    height: 100%;
+    width: 60%;
+    margin-left: auto;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    font-size: 1rem;
+    color: white;
+  }
+
+  .seller-follow {
+    width: 40%;
+    height: 20%;
+    border-radius: 5rem;
+    background-color: #fff;
+    color: black;
+    border-radius: 2rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    font-size: 1vw;
+    box-shadow: 5px 5px 10px rgba(0, 0, 0, 0.5);
+    margin-bottom: 0.3rem;
+    cursor: pointer;
+  }
+
+  .seller-follow:hover {
+    background-color: black;
+    color: white;
+  }
+}
