--- test.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!-- Bootstrap -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
-    <!-- CSS -->
-    <link rel="stylesheet" href="static/css/myreview.css" />
-    <title>ChocoTheCoo</title>
-</head>
-
-<!-- header -->
-<header></header>
-
-<!-- body -->
-
-<body>
-
-</body>
-<!-- footer -->
-<footer></footer>
-<script src="/static/js/api.js" type="module"></script>
-<script src="/static/js/loader.js" type="module"></script>
-
-</html>
